LPS-61524 Rename portal-cache-ehcache-single to portal-cache-single, as it has nothing to do with ehcache. It works at portal-cache level.
