--- modules/portal-cache/portal-cache-ehcache-single/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Portal Cache Ehcache Single
-Bundle-SymbolicName: com.liferay.portal.cache.ehcache.single
-Bundle-Version: 1.0.0
-Include-Resource: classes
