+++ modules/portal-cache/portal-cache-single/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Cache Single
+Bundle-SymbolicName: com.liferay.portal.cache.single
+Bundle-Version: 1.0.0
+Include-Resource: classes
