+++ modules/portal-cache/portal-cache-single/src/main/java/com/liferay/portal/cache/single/internal/distribution/SinglePortalCacheReplicatorFactory.java
@@ -12,10 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache.single.internal.bootstrap;
+package com.liferay.portal.cache.single.internal.distribution;
 
-import com.liferay.portal.cache.PortalCacheBootstrapLoader;
-import com.liferay.portal.cache.PortalCacheBootstrapLoaderFactory;
+import com.liferay.portal.cache.PortalCacheReplicator;
+import com.liferay.portal.cache.PortalCacheReplicatorFactory;
+
+import java.io.Serializable;
 
 import java.util.Properties;
 
@@ -24,12 +26,14 @@ import org.osgi.service.component.annotations.Component;
 /**
  * @author Tina Tian
  */
-@Component(immediate = true, service = PortalCacheBootstrapLoaderFactory.class)
-public class SinglePortalCacheBootstrapLoaderFactory
-	implements PortalCacheBootstrapLoaderFactory {
+@Component(immediate = true, service = PortalCacheReplicatorFactory.class)
+public class SinglePortalCacheReplicatorFactory
+	implements PortalCacheReplicatorFactory {
 
 	@Override
-	public PortalCacheBootstrapLoader create(Properties properties) {
+	public <K extends Serializable, V extends Serializable>
+		PortalCacheReplicator<K, V> create(Properties properties) {
+
 		return null;
 	}
